pyproject.toml
[project]
name = "orders-master-infoprex"
version = "1.0.0"
description = "Orders Master Infoprex - Sell Out & Orders Consolidation"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30,<1.40",
    "pandas>=2.0,<3.0",
    "openpyxl>=3.1,<4.0",
    "pydantic>=2.0,<3.0",
    "python-dotenv>=1.0,<2.0",
    "pyyaml>=6.0,<7.0",
    "python-dateutil>=2.8,<3.0",
    "pandera>=0.18,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0,<9.0",
    "pytest-cov>=4.0,<6.0",
    "pytest-benchmark>=4.0,<5.0",
    "ruff>=0.5,<1.0",
    "black>=24.0,<25.0",
    "mypy>=1.0,<2.0",
    "pre-commit>=3.0,<4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "PL", "UP", "W"]
ignore = ["PLR2004", "E501", "PLR0912", "PLR0915"]

[tool.ruff.lint.isort]
known-first-party = ["orders_master", "ui"]

[tool.ruff.format]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
strict = true
python_version = "3.11"

[[tool.mypy.overrides]]
module = "pandas.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "streamlit.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "openpyxl.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pandera.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
